third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/module/import-meta/import-meta-url-expected.txt

This is a testharness.js-based test.
[FAIL] import.meta.url when importing the module with different fragments
  assert_equals: expected "http://web-platform.test:8001/html/semantics/scripting-1/the-script-element/module/import-meta/import-meta-root.js#1" but got "http://web-platform.test:8001/html/semantics/scripting-1/the-script-element/module/import-meta/import-meta-root.js"
Harness: the test ran to completion.
